added the Tests for the Services & Integration test for the authentication

+++ src/main/java/com/track/share/user/UserServiceImpl.java
@@ -1,25 +1,26 @@
 package com.track.share.user;
 
 import java.util.List;
-import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.track.share.exceptions.NotFoundException;
+import com.track.share.exceptions.UsernameUnavailableException;
 
+import lombok.AllArgsConstructor;
 
 @Service
+@AllArgsConstructor
 class UserServiceImpl implements UserService {
 
-	@Autowired
 	private UserRepository userRepository;
 
 	@Override
 	public Users createUser(Users user) {
-		Optional<Users> u = userRepository.findByEmail(user.getEmail());
-		if(!u.isEmpty())
-			return null;
+		if (userRepository.existsByEmail(user.getEmail())) {
+			throw new UsernameUnavailableException("email is Unavailable, please try with another email");
+		}
 		user.setStatus(true);
 		user.setPassword(new BCryptPasswordEncoder().encode(user.getPassword()));
 		return userRepository.save(user);
@@ -32,11 +33,15 @@ class UserServiceImpl implements UserService {
 
 	@Override
 	public Users getUser(String username) {
-		return userRepository.findByEmail(username).orElse(null);
+		return userRepository.findByEmail(username)
+				.orElseThrow(() -> new NotFoundException("user not found with username " + username));
 	}
 
 	@Override
 	public void removeUser(Long userId) {
+		if (!userRepository.existsById(userId)) {
+			throw new NotFoundException("user not found with userId " + userId);
+		}
 		userRepository.deleteById(userId);
 	}
